Begin moving rooms stuff to their own view file
